Removed the majority of ListPages CSS, it's unwanted
The ListPages extension isn't coming back anytime soon, if ever. Of the
current social tools, the following use the listpages-* CSS classes due to
legacy reasons, but they (should) provide their own styling:
* BlogPage
* LinkFilter
* SportsTeams

This way we don't load a lot of unnecessary CSS in case if a wiki isn't
even using social tools in the first place, let alone one (or more) of the
aforementioned extensions.

The rating-total CSS class was also originally used by ListPages (despite
not being prefixed with the listpages- prefix), but is nowadays used by
VoteNY.

Change-Id: I0a35748a062e2c640c4d1d09b43b1ec9f428fd68