Some cleanup, replace small images used for styling with pure CSS
This adds two base CSS classes for some skin components in Nimbus:
 - mw-skin-nimbus-button
 - mw-skin-nimbus-actiontab

This patch does a general cleanup of these styles to reduce
CSS code duplication, as well as rewrite the styles so
that we no longer need images (now done with CSS). There's only
1 exception which is the right arrow icon, which will probably
be replaced with the OOUI right arrow icon in the future.

These new styles are placed under resources/styles. In the future,
there'll be a patch that will move the rest of JS/CSS files (from /nimbus
and /skinStyles) into the /resources folder to follow code conventions.

Bug: T155280
Change-Id: Icd8627ee9d5557845bcf61da6de050a7fa63374a