Refactor the content action tab generation to use more core methods so that certain hooks run etc.
* Gets rid of the now unnecessary and awfully outdated tabAction()
* More RequestContext in a few places
* Added the 'id' attribute to the generated HTML output for both "regular" tabs as the ones grouped under "More actions"
** This is needed by the ProblemReports extension, which looks for a certain ID (#ca-report-problem on non-Monaco, #fe_report_link on Monaco) and launches the dialog once that link is clicked
* Swapped strpos() to preg_match() in the foreach ( $leftLinks as $key => $val ) {} loop to ensure that "Discussion" tabs also properly show up as red when the talk page doesn't exist
** The old strpos() could would work when $val['class'] = 'selected new'; but fail when $val['class'] = 'new';

The usage of ReflectionMethod here to access protected/private SkinTemplate methods is somewhat nasty, but it works as intended and doesn't force us to copy those methods over.

As a nice side-effect of this refactoring, the "More actions" tab is visible on nonexistent pages, allowing admins to access the protection/unprotection menu from there and allowing all users to easily access Special:WhatLinksHere for the page.

Change-Id: I52919b518cf2fb4a7e907e624665ce15e1786e46